+++ packages/www/lib/mdx.d.ts
@@ -1,8 +1,0 @@
-/// <reference types="react/canary" />
-
-declare module "*.mdx" {
-  import type { ComponentType } from "react";
-
-  const MDXComponent: ComponentType;
-  export default MDXComponent;
-}